Specify type of water demand fields in `SBM` And move function `update_water_demand(sbm::SBM)` to sbm.jl.
