+++ src/Wflow.jl
@@ -115,6 +115,7 @@ Base.show(io::IO, m::Model) = print(io, "model of type ", typeof(m))
 
 include("horizontal_process.jl")
 include("hbv.jl")
+include("water_demand.jl")
 include("sbm.jl")
 include("flextopo.jl")
 include("flow.jl")
@@ -132,7 +133,6 @@ include("sbm_gwf_model.jl")
 include("utils.jl")
 include("bmi.jl")
 include("subdomains.jl")
-include("water_demand.jl")
 include("logging.jl")
 include("states.jl")
